+++ indico/modules/events/registration/models/forms.py
@@ -1,5 +1,5 @@
 # This file is part of Indico.
-# Copyright (C) 2002 - 2020 CERN
+# Copyright (C) 2002 - 2021 CERN
 #
 # Indico is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see the
@@ -29,10 +29,11 @@ from indico.util.struct.enum import RichIntEnum
 
 
 class ModificationMode(RichIntEnum):
-    __titles__ = [None, L_('Until modification deadline'), L_('Until payment'), L_('Never')]
+    __titles__ = [None, L_('Until modification deadline'), L_('Until payment'), L_('Never'), L_('Until approved')]
     allowed_always = 1
     allowed_until_payment = 2
     not_allowed = 3
+    allowed_until_approved = 4
 
 
 class RegistrationForm(db.Model):
@@ -402,6 +403,8 @@ class RegistrationForm(db.Model):
             return False
         elif self.modification_mode == ModificationMode.allowed_always:
             return True
+        elif self.modification_mode == ModificationMode.allowed_until_approved:
+            return registration.state == RegistrationState.pending
         elif self.modification_mode == ModificationMode.allowed_until_payment:
             return not registration.is_paid
         else: