+++ indico/util/mdx_latex.py
@@ -1,5 +1,5 @@
 # This file is part of Indico.
-# Copyright (C) 2002 - 2020 CERN
+# Copyright (C) 2002 - 2021 CERN
 #
 # Indico is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see the
@@ -153,6 +153,9 @@ def unescape_html_entities(text):
 
 
 def latex_escape(text, ignore_math=True, ignore_braces=False):
+    if text is None:
+        return ''
+
     chars = {
         "#": r"\#",
         "$": r"\$",